backend/requirements.txt
fastapi>=0.116.1,<1.0
uvicorn>=0.35.0,<1.0
sqlalchemy>=2.0.41,<3.0
python-dotenv>=1.1.0,<2.0
pydantic>=2.12.0,<3.0
pydantic-settings>=2.13.0,<3.0
python-jose[cryptography]>=3.5.0,<4.0
cryptography>=45.0.5,<46.0
email-validator>=2.2.0,<3.0
python-multipart>=0.0.20,<0.1
